chore(nexus): clean up require codepath (#6616)
