+++ wandb/util.py
@@ -1837,3 +1837,23 @@ def sample_with_exponential_decay_weights(
     sampled_keys = keys_array[sampled_indices].tolist() if keys else None
 
     return sampled_xs, sampled_ys, sampled_keys
+
+
+def get_core_path() -> str:
+    core_path: str = os.environ.get("_WANDB_NEXUS_PATH", "")
+    wandb_core = get_module("wandb_core")
+    if not core_path and wandb_core:
+        _check_wandb_core_version_compatibility(wandb_core.__version__)
+        core_path = wandb_core.get_nexus_path()
+    return core_path
+
+
+def _check_wandb_core_version_compatibility(core_version: str) -> None:
+    """Checks if the installed wandb-core version is compatible with the wandb version."""
+    from pkg_resources import parse_version
+
+    if parse_version(core_version) < parse_version(wandb._minimum_core_version):
+        raise ImportError(
+            f"Requires wandb-core version {wandb._minimum_core_version} or later, "
+            f"but you have {core_version}. Run `pip install --upgrade wandb[core]` to upgrade."
+        )
